// src/generated/models/SDDestinationModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface SDDestination {
  Code: string;
  Name?: string;
  DocEntry: number;
  Canceled?: string;
  Object?: string;
  LogInst?: number;
  UserSign?: number;
  Transfered?: string;
  CreateDate?: string;
  CreateTime?: number;
  UpdateDate?: string;
  UpdateTime?: number;
  DataSource?: string;
  NaturalPer?: string;
  DPPStatus?: string;
  EncryptIV?: string;
  U_Active?: string;
  U_Description?: string;
  U_DeliveryType?: string;
  U_Category?: string;
  U_BusinessType?: string;
  U_MainDst?: string;
}